# backend/contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Tingyue Backend API",
    "version": "1.0.0",
    "description": "Authoritative API contract api-contract-v1.0.0. JSON fields use camelCase and timestamps use UTC RFC 3339. Clients branch on stable error.code values, never human messages."
  },
  "servers": [
    { "url": "http://127.0.0.1:3100", "description": "Local development" },
    { "url": "https://api.test.example.invalid", "description": "Test placeholder; replace after domain approval" }
  ],
...
  "security": [{ "bearerAuth": [] }],
  "paths": {
...
    "/api/v1/admin/ingestion-batches": {
      "post": {
        "operationId": "createIngestionBatch",
        "tags": ["Admin ingestion"],
        "description": "Creates only control-plane metadata. File bytes are uploaded directly to object storage.",
        "parameters": [{ "$ref": "#/components/parameters/IdempotencyKey" }],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "./schemas/api-v1.schema.json#/$defs/IngestionBatchCreateRequest" } } }
        },
        "responses": {
          "201": {
            "description": "Batch and declared items created.",
            "headers": { "X-Request-Id": { "$ref": "#/components/headers/RequestId" } },
            "content": { "application/json": { "schema": { "$ref": "./schemas/api-v1.schema.json#/$defs/IngestionBatchResponse" } } }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "409": { "$ref": "#/components/responses/Conflict" }
        }
      }
    },
    "/api/v1/admin/ingestion-batches/{batchId}": {
      "get": {
        "operationId": "getIngestionBatch",
        "tags": ["Admin ingestion"],
        "parameters": [{ "$ref": "#/components/parameters/BatchId" }],
        "responses": {
          "200": {
            "description": "Batch, item and processing status without presigned URLs or sensitive payloads.",
            "headers": { "X-Request-Id": { "$ref": "#/components/headers/RequestId" } },
            "content": { "application/json": { "schema": { "$ref": "./schemas/api-v1.schema.json#/$defs/IngestionBatchResponse" } } }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
    "/api/v1/admin/upload-sessions": {
      "post": {
        "operationId": "createUploadSession",
        "tags": ["Admin ingestion"],
        "parameters": [{ "$ref": "#/components/parameters/IdempotencyKey" }],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "./schemas/api-v1.schema.json#/$defs/UploadSessionCreateRequest" } } }
        },
        "responses": {
          "201": {
            "description": "Short-lived multipart upload session. URLs are never written to ordinary logs.",
            "headers": { "X-Request-Id": { "$ref": "#/components/headers/RequestId" } },
            "content": { "application/json": { "schema": { "$ref": "./schemas/api-v1.schema.json#/$defs/UploadSessionResponse" } } }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "409": { "$ref": "#/components/responses/Conflict" }
        }
      }
    },
    "/api/v1/admin/upload-sessions/{uploadSessionId}/complete": {
      "post": {
        "operationId": "completeUploadSession",
        "tags": ["Admin ingestion"],
        "parameters": [
          { "$ref": "#/components/parameters/UploadSessionId" },
          { "$ref": "#/components/parameters/IdempotencyKey" }
        ],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "./schemas/api-v1.schema.json#/$defs/UploadSessionCompleteRequest" } } }
        },
        "responses": {
          "200": { "$ref": "#/components/responses/AdminAction" },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "409": { "$ref": "#/components/responses/Conflict" },
          "410": { "$ref": "#/components/responses/Gone" },
          "422": { "$ref": "#/components/responses/Unprocessable" }
        }
      }
    },
    "/api/v1/admin/ingestion-batches/{batchId}/finalize": {
      "post": {
        "operationId": "finalizeIngestionBatch",
        "tags": ["Admin ingestion"],
        "parameters": [
          { "$ref": "#/components/parameters/BatchId" },
          { "$ref": "#/components/parameters/IdempotencyKey" }
        ],
        "responses": {
          "202": { "$ref": "#/components/responses/AdminAction" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "409": { "$ref": "#/components/responses/Conflict" },
          "422": { "$ref": "#/components/responses/Unprocessable" }
        }
      }
    },
    "/api/v1/admin/ingestion-batches/{batchId}/cancel": {
      "post": {
        "operationId": "cancelIngestionBatch",
        "tags": ["Admin ingestion"],
        "parameters": [
          { "$ref": "#/components/parameters/BatchId" },
          { "$ref": "#/components/parameters/IdempotencyKey" }
        ],
        "responses": {
          "200": { "$ref": "#/components/responses/AdminAction" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "409": { "$ref": "#/components/responses/Conflict" }
        }
      }
    },
    "/api/v1/admin/ingestion-jobs/{jobId}/retry": {
      "post": {
        "operationId": "retryIngestionJob",
        "tags": ["Admin ingestion"],
        "parameters": [
          { "$ref": "#/components/parameters/JobId" },
          { "$ref": "#/components/parameters/IdempotencyKey" }
        ],
        "responses": {
          "202": { "$ref": "#/components/responses/AdminAction" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "409": { "$ref": "#/components/responses/Conflict" }
        }
      }
    },
    "/api/v1/admin/content-versions/{contentVersionId}/publish": {
      "post": {
        "operationId": "publishContentVersion",
        "tags": ["Admin ingestion"],
        "description": "Atomically changes the published manifest only after all technical and human review gates are satisfied.",
        "parameters": [
          { "$ref": "#/components/parameters/ContentVersionId" },
          { "$ref": "#/components/parameters/IdempotencyKey" }
        ],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "./schemas/api-v1.schema.json#/$defs/PublishContentVersionRequest" } } }
        },
        "responses": {
          "200": { "$ref": "#/components/responses/AdminAction" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "409": { "$ref": "#/components/responses/Conflict" }
        }
      }
    },
    "/api/v1/admin/content-versions/{contentVersionId}/rollback": {
      "post": {
        "operationId": "rollbackContentVersion",
        "tags": ["Admin ingestion"],
        "parameters": [
          { "$ref": "#/components/parameters/ContentVersionId" },
          { "$ref": "#/components/parameters/IdempotencyKey" }
        ],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "./schemas/api-v1.schema.json#/$defs/RollbackContentVersionRequest" } } }
        },
        "responses": {
          "200": { "$ref": "#/components/responses/AdminAction" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "409": { "$ref": "#/components/responses/Conflict" }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Short-lived access token. Tokens must not be logged."
      }
    },
    "headers": {
      "RequestId": {
        "description": "Opaque request correlation identifier.",
        "required": true,
        "schema": { "type": "string", "minLength": 8, "maxLength": 128 }
      }
    },
    "parameters": {
      "IdempotencyKey": {
        "name": "Idempotency-Key",
        "in": "header",
        "required": true,
        "description": "Opaque key unique to one logical mutation. Reusing it with a different request body returns IDEMPOTENCY_KEY_REUSED.",
        "schema": { "type": "string", "minLength": 8, "maxLength": 128 }
      },
...
      "BatchId": { "name": "batchId", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } },
      "UploadSessionId": { "name": "uploadSessionId", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } },
      "JobId": { "name": "jobId", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } },
      "ContentVersionId": { "name": "contentVersionId", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
    },
    "responses": {
      "BadRequest": {
        "description": "Malformed request, invalid ranking period or invalid filter.",
        "headers": { "X-Request-Id": { "$ref": "#/components/headers/RequestId" } },
        "content": { "application/json": { "schema": { "$ref": "./schemas/api-v1.schema.json#/$defs/ErrorResponse" } } }
      },
      "Unauthorized": {
        "description": "Authentication failed, expired or was revoked.",
        "headers": { "X-Request-Id": { "$ref": "#/components/headers/RequestId" } },
        "content": { "application/json": { "schema": { "$ref": "./schemas/api-v1.schema.json#/$defs/ErrorResponse" } } }
      },
      "Forbidden": {
        "description": "Authenticated identity lacks the required permission or content access.",
        "headers": { "X-Request-Id": { "$ref": "#/components/headers/RequestId" } },
        "content": { "application/json": { "schema": { "$ref": "./schemas/api-v1.schema.json#/$defs/ErrorResponse" } } }
      },
      "NotFound": {
        "description": "Business resource not found or not visible.",
        "headers": { "X-Request-Id": { "$ref": "#/components/headers/RequestId" } },
        "content": { "application/json": { "schema": { "$ref": "./schemas/api-v1.schema.json#/$defs/ErrorResponse" } } }
      },
      "Conflict": {
        "description": "State, version or idempotency conflict.",
        "headers": { "X-Request-Id": { "$ref": "#/components/headers/RequestId" } },
        "content": { "application/json": { "schema": { "$ref": "./schemas/api-v1.schema.json#/$defs/ErrorResponse" } } }
      },
      "Gone": {
        "description": "The upload session has expired.",
        "headers": { "X-Request-Id": { "$ref": "#/components/headers/RequestId" } },
        "content": { "application/json": { "schema": { "$ref": "./schemas/api-v1.schema.json#/$defs/ErrorResponse" } } }
      },
      "Unprocessable": {
        "description": "Checksum, subtitle or processing validation failed.",
        "headers": { "X-Request-Id": { "$ref": "#/components/headers/RequestId" } },
        "content": { "application/json": { "schema": { "$ref": "./schemas/api-v1.schema.json#/$defs/ErrorResponse" } } }
      },
...
      "AdminAction": {
        "description": "Idempotent administrative action accepted or completed.",
        "headers": { "X-Request-Id": { "$ref": "#/components/headers/RequestId" } },
        "content": { "application/json": { "schema": { "$ref": "./schemas/api-v1.schema.json#/$defs/AdminActionResponse" } } }
      }
    }
  }
}